saved user registration details in user preferences

+++ app/src/main/java/com/kabbodev/emaishapay/data/models/AuthenticationResponse.kt
@@ -35,6 +35,6 @@ class AuthenticationResponse(
         val phoneNumber: String?,
         @SerializedName("balance")
         @Expose
-        var balance: Double?
+        var balance: Long?
 )
